Fix screen orientation as portrait
Android restarts activity by default when device configuration (e.g.,
screen orientation) changes. This requires us to reconnect to the
camera which is costly and causes race conditions.

Fix the screen orientation for our application to portait. This
ensures that we won't be restarted when the user turns their device,
and portrait seems to work best for my camera preview (it is so
abnormally wiiiiiiide in landscape mode).