Better camera management
onResume() and onPause() are not the best places to manage camera,
and unlock() and reconnect() methods are not the ones we should use.

Keeping a reference to the camera prevents other apps from using it
(e.g., the default Camera app does not even start). We should release
the camera when we are stopped, i.e. move to background and cannot
effectively make use of the camera.