Track device rotation
We're now using SensorManager to track the rotation of the device
and adjust the camera viewpoint accordingly. All of this results
in the cube being shown from different angles, correlated with
the device rotation.

The rendering is updated as well. It turned out that we need to
enable face culling manually. Without it the cube was rendered
incorrectly.

The rotation sensor on my device seems to be very noisy. The cube
wiggles around even when the device lays stationary on the table.
We'll probably need to filter out the rotation events, or maybe
coordinate rotation with the data from the camera. This sounds hard.