Improve rotation stability
Increase sampling rate as this seems to improve the output of the
sensor. Also add some simple smoothing of the sensor data so that
the changes are not propagated instantly. This gives some minor
improvements in rotation stability.

Maybe including positional data will also have some effect on the
visible image. Accelerometer integration is next.