Add AstroLib 1.1.5ws
I was tired of trying to wrap my head around various astronomical
computations. While it's certainly an interesting thing to research
I have found that it blocks me from actually building a fucking app
so I decided to ditch that effort and just use a library.

I have no clue about how libraries work in Android, but I really
dislike putting binaries into source code repository. And this
library is not on Maven or whatever. So it is added as a source
module and will be compiled as a part of this project.

There was erroneous UTF-8 code sequence in the source. There are no
source code changes other than that and line ending conversion.