Computing planet locations
This adds routines for computing planet locations using AstroLib. The
library has a bit wonky interface, but that's the best that we have.
At least it seems to do the calculations right. The results are cross-
referenced with NASA JPL HORIZONS system which is more or less trusted.
Well, I believe in them. These people were launching people is space
when I was not even born.

We go for geocentric equatorial polar coordinates (aka RA+Dec) because
that's the coordinate system used by most star catalogs and we'd better
have stars and planets in the same format so that they can be rendered
together.