Stellar data
Here we pull in some data about stars. I have looked in the Internet
and the HYG project seems to fit best, it provides modern HIPPARCOS
numbers for stars, Bayer designations, location data, and spectra.
It should be enough for my goals. Thank you, Dave, for compiling
this data set.

Full data set is about 14 MB compressed and 34 MB uncompressed in CSV.
That's way too huge and we don't need every single star. Only stars
visible to the naked eye (magnitude 6 and below) should be enough.

I have decided to use Perl for pulling and preprocessing the data
because it seems to fit nicely into Perl's domain. We pull the catalog,
filter out interesting stars and columns, and print it out as XML.

Yes, XML. Java cannot into static data sets so we have to use a bit
roundabout way. The data set is included as an XML resource which
we manually parse during application startup. Well... good enough.