Rendering the stellar dome
This is the first attempt at actually rendering stars. Throw away the
Box, draw Stars instead. The rendering is pretty simple, but it took
me a while to remember all the linear math involved with 3D matrices
and implement a billboarding transformation.

It kinda works, but lags tremendously on my device, pumping out 10 FPS
or something like that. I've tried profiling it and it seems that we
do too many glDrawElements() calls. Maybe we should try computing the
vertex coordinates in Java code. Or try culling the stars so that we
send only visible ones to OpenGL, not all five thousands of them.
Every frame.