Precompute star quad positions
It turned out that setting model-view-projection matrices and calling
glDrawElements() is *very* costly. So rendering was very inefficient.
At least on my device.

If we don't require the quads to be billboarded ideally (so that they
are always exactly horizontal) then we can precompute their positions
and just feed a single matrix and several thousand triangles in a
single call. That's way better for OpenGL performance.