Use textures for star quads
Now we're able to draw textured stars. This could be used to make them
look actually round. (Not much at this point.) Later we could multiply
the texture with the star color.

This has required a bunch of new data buffers and some changes in the
shaders. But nothing critical. We'll refactor the code later.