Simplify buffer allocation
It's kinda cute to avoid unnecessary memory reallocations when
computing new star data, but that also makes the code complicated.
Drop all of this and simply allocate a new buffer when we need it.
It's not like we will be updating this data really often. Even if
we end up updating it, if this becomes a problem, we can introduce
some special in-place update method which tweaks only vertex data
to account for star (or rather, Earth-based observer) movement.